.github/workflows/workflow_gen/pyproject.toml
[project]
name = "workflow-gen"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = ["ruamel.yaml>=0.18"]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
